refactor: 기상 그래프 API 응답값 double 형으로 수정

+++ backend/src/main/java/com/imsnacks/Nyeoreumnagi/weather/dto/response/GetWeatherGraphResponse.java
@@ -12,7 +12,7 @@ public record GetWeatherGraphResponse (
 ){
     public record ValuePerTime(
             String name,
-            int value
+            double value
     ){
     }
 }
